+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [9.113.0] - 2020-04-07
+
+### Added
+
+- `fit` prop to `PageBlock` and `Box` components.
+
 ## [9.112.28] - 2020-04-02
 
 ### Fixed